@@ -1,19 +1,17 @@
 plugins {
-	id "org.springframework.boot" version "2.7.10"
-	id "io.spring.dependency-management" version "1.0.11.RELEASE"
+	id "org.springframework.boot" version "3.0.5"
+	id "io.spring.dependency-management" version "1.1.0"
 	id "java"
 }
 
 group = "com.example"
 version = "0.0.1-SNAPSHOT"
-sourceCompatibility = "11"
+sourceCompatibility = "17"
 
 repositories {
 	mavenCentral()
 	maven { url "https://repo.spring.io/release" }
 }
-
-ext["spring-security.version"] = "5.8.2"
 
 dependencies {
 	implementation "org.springframework.boot:spring-boot-starter-data-jpa"
